Move PreprocessFile function to ebpf package
