+++ pkg/collector/corechecks/ebpf/probe/oom_kill.go
@@ -7,7 +7,6 @@ import (
 	"unsafe"
 
 	"github.com/DataDog/datadog-agent/pkg/ebpf"
-	"github.com/DataDog/datadog-agent/pkg/ebpf/bytecode"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
 	bpflib "github.com/iovisor/gobpf/bcc"
@@ -25,7 +24,7 @@ type OOMKillProbe struct {
 }
 
 func NewOOMKillProbe(cfg *ebpf.Config) (*OOMKillProbe, error) {
-	source, err := bytecode.PreprocessFile(cfg.BPFDir, "oom-kill-kern.c")
+	source, err := ebpf.PreprocessFile(cfg.BPFDir, "oom-kill-kern.c")
 	if err != nil {
 		return nil, fmt.Errorf("Couldn’t process headers for asset “oom-kill-kern.c”: %v", err)
 	}
